AlphaBetaDecider: Fix logic for max/min functions
Apparently I was rather confused when writing the original max/min
functions, because most of the PriorityQueue stuff was plain wrong.
Also, I was calling the wrong functions in the main driver method.
You are supposed to minimize the move you check, bc. it's the other
player's turn. All in all, things should work better now.